Extract `CursorContext` to its own file
